Web App + MongoDB and Web App + SQL

+++ samples/web-app-cosmosdb-mongodb-api/python/terraform/deploy.sh
@@ -1,8 +1,12 @@
 #!/bin/bash
 
 # Variables
+PREFIX='paolo'
+SUFFIX='test'
+LOCATION='westeurope'
 CURRENT_DIR="$(cd "$(dirname "$0")" && pwd)"
 ZIPFILE="planner_website.zip"
+ENVIRONMENT=$(az account show --query environmentName --output tsv)
 
 # Change the current directory to the script's directory
 cd "$CURRENT_DIR" || exit
@@ -10,16 +14,24 @@ cd "$CURRENT_DIR" || exit
 # Start azure CLI local mode session
 # azlocal start_interception
 
-# Delete any existing terraform state and plan files
-rm -f terraform.tfstate terraform.tfstate.backup tfplan
-
 # Run terraform init and apply
+if [[ $ENVIRONMENT == "LocalStack" ]]; then
+	echo "Using tflocal for LocalStack emulator environment."
+	TERRAFORM_CMD="tflocal"
+else
+	echo "Using standard terraform for AzureCloud environment."
+	TERRAFORM_CMD="terraform"
+fi
+
 echo "Initializing Terraform..."
-tflocal init -upgrade
+$TERRAFORM_CMD init -upgrade
 
 # Run terraform plan and check for errors
 echo "Planning Terraform deployment..."
-tflocal plan -out=tfplan
+$TERRAFORM_CMD plan -out=tfplan \
+	-var="prefix=$PREFIX" \
+	-var="suffix=$SUFFIX" \
+	-var="location=$LOCATION"
 
 if [[ $? != 0 ]]; then
 		echo "Terraform plan failed. Exiting."
@@ -56,12 +68,27 @@ echo "Creating zip package of the web app..."
 zip -r "$ZIPFILE" app.py cosmosdb.py static templates requirements.txt
 
 # Deploy the web app
-azlocal webapp deploy \
-  --resource-group "$RESOURCE_GROUP_NAME" \
-  --name "$WEB_APP_NAME" \
-  --src-path planner_website.zip \
-  --type zip \
-  --async true
+echo "Deploying web app [$WEB_APP_NAME] with zip file [$ZIPFILE]..."
+if [[ $ENVIRONMENT == "LocalStack" ]]; then
+	echo "Using azlocal webapp deploy command for LocalStack emulator environment."
+	azlocal webapp deploy \
+		--resource-group "$RESOURCE_GROUP_NAME" \
+		--name "$WEB_APP_NAME" \
+		--src-path "$ZIPFILE" \
+		--type zip \
+		--async true
+else
+	echo "Using standard az webapp deploy command for AzureCloud environment."
+	az webapp deploy \
+		--resource-group "$RESOURCE_GROUP_NAME" \
+		--name "$WEB_APP_NAME" \
+		--src-path "$ZIPFILE" \
+		--type zip \
+		--async true
+fi
 
 # Remove the zip package of the web app
-rm "$ZIPFILE"
+if [ -f "$ZIPFILE" ]; then
+	rm "$ZIPFILE"
+fi
+
